+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- More logging for grist cost generator to help identify bottlenecks
+
 ## [1.20.1-1.12.0.0] - 2024-04-13
 
 ### Added